requirements.txt
# FastAPI本体とサーバー
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# テスト
pytest>=8.2.0
httpx>=0.27.0

# 環境変数
python-dotenv>=1.0.1

# Azure SDK
azure-identity>=1.17.0
azure-core>=1.30.0
azure-storage-blob>=12.21.0
azure-search-documents>=11.6.0

# OpenAI (Azure経由)
openai>=1.35.0

# その他ユーティリティ（任意）
pydantic>=2.8.0
python-multipart
isort

